add progress's example
